Flytt kode som har mer med generering av vilkårsvurdering enn vilkår å gjøre fra vilkårService til VilkårsvurderingForNyBehandlingService
